split exceptions.py to 2 files
