+++ main.py
@@ -2,8 +2,8 @@ import asyncio
 
 from src.config.config import Config
 from src.constants.constants import LPPARSER_LOGOTYPE
-from src.exceptions.exceptions import (
-    ApplicationException,
+from src.exceptions.base import ApplicationException
+from src.exceptions.config import (
     ConfigException,
     ConfigNotFoundError,
     IncorrectOffsetError,
